requirements.txt
aiofiles==23.1.0
aiogithub @ git+https://github.com/reupen/aiogithub.git@7dab796756b36f6c2535e3fec2e5318853357d10
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
black==23.1.0
braceexpand==0.1.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
discord.py==2.2.2
frozenlist==1.3.3
humanize==4.6.0
idna==3.4
import-expression==1.1.4
isort==5.12.0
jishaku==2.5.1
line-profiler==4.0.3
LinkHeader==0.4.3
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
psutil==5.9.4
pycparser==2.21
pygit2==1.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
tabulate==0.9.0
typing_extensions==4.5.0
uritemplate==3.0.1
yarl==1.8.2
